requirements.txt
torch~=1.10.0
torchvision~=0.11.1
tqdm~=4.62.3
xgboost~=1.5.0
pillow~=8.4.0
numpy~=1.21.4
timm~=0.4.12
sklearn~=0.0
scikit-learn~=1.0.1